Major refactor, preparing for v2 session encoding format.
