+++ lib/global_session/session/abstract.rb
@@ -1,0 +1,5 @@
+module GlobalSession::Session
+  class Abstract
+    attr_reader :id, :authority, :created_at, :expired_at, :directory
+  end
+end
